Move roomCard to separate file
